Allow passing custom tag label to metrics (#12)
This is useful if users for example want to set up alerts based some
custom grouping.
I originally wanted to go with totally custom tags but after
implementing it, it turned out that rust-prometheus does not support
this well. We would always need to provide all labels for every metric.
So if one address has the custom label "a=a" and another the custom label
"b=b" then we need to set both "a" and "b" for both metrics. We could
implement that but it feels annoying enough to do this easier version
instead.